refactor: externalize settings to YAML and split model instantiation
- Moved assistant configuration (domains, logs dir, assistant mode) to configs/config.yaml
- Updated settings.py to load constants from YAML and environment
- Created model_factory.py to handle ChatOllama model instantiation
