+++ setup_and_pull.sh
