--- src/os_assistant/config/__init__.py
@@ -1,1 +1,0 @@
-# This file marks the config directory as a Python package.
