+++ src/os_assistant/configs/config.yaml
@@ -1,0 +1,16 @@
+DOMAINS:
+  - file_system
+  - users
+  - packages
+  - network
+
+LOGS_DIR: "domain_logs"
+
+TEMPERATURE: 0
+
+# Assistant mode:
+# 0 = Both code tool + RAG
+# 1 = Code tool only
+# 2 = RAG only
+# 3 = Basic model only
+ASSISTANT_MODE: 1
