+++ src/os_assistant/core/state.py
